% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_calc.R
\name{my_calc}
\alias{my_calc}
\title{A function to calculate a numeric vectors or metrices.}
\usage{
my_calc(x, y, fun)
}
\arguments{
\item{x}{A numeric vector or metrices}

\item{y}{A numeric vector or metrices}

\item{fun}{A character values}
}
\description{
This function allows you to calculate two vectors or metrices.
fun parameter values must be among “sum”, “subtract”, “divide”, “multiply”
}
\examples{
my_calc(c(1:10),c(2,5), "sum")
my_calc(c(1:10),c(2,5), "subtract")
my_calc(c(1:10),c(2,5), "divide")
my_calc(c(1:10),c(2,5), "multiply")
}
