% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_multiplier.R
\name{my_multiplier}
\alias{my_multiplier}
\title{A number-checking function}
\usage{
my_multiplier(value)
}
\arguments{
\item{value}{A number}
}
\description{
This function allows you to test whether a number falls into the
Goldilocks range (5-9) or not
}
\examples{
my_multiplier(2)
my_multiplier(6)
my_multiplier(12)
}
